Backport empty statement fix
Split Ch13 tests from Chapter12Test.java
Handle forward refs never defined.
