+++ chapter14/src/test/java/com/seaofnodes/simple/Chapter12Test.java
@@ -59,116 +59,4 @@ return x+1==x;
         assertEquals("return ((flt)arg==(ToFloat+1.0));", stop.toString());
         assertEquals(0L, Evaluator.evaluate(stop, 1));
     }
-
-
-    @Test
-    public void testLinkedList0() {
-        Parser parser = new Parser(
-"""
-struct LLI { LLI? next; int i; }
-LLI? head = null;
-while( arg ) {
-    LLI x = new LLI;
-    x.next = head;
-    x.i = arg;
-    head = x;
-    arg = arg-1;
-}
-return head.next.i;
-""");
-        try { parser.parse(true).iterate(true); fail(); }
-        catch( Exception e ) { assertEquals("Might be null accessing 'i'",e.getMessage()); }
-    }
-
-    @Test
-    public void testLinkedList1() {
-        Parser parser = new Parser(
-"""
-struct LLI { LLI? next; int i; }
-LLI? head = null;
-while( arg ) {
-    LLI x = new LLI;
-    x.next = head;
-    x.i = arg;
-    head = x;
-    arg = arg-1;
-}
-if( !head ) return 0;
-LLI? next = head.next;
-if( next==null ) return 1;
-return next.i;
-""");
-        StopNode stop = parser.parse(false).iterate(false);
-        assertEquals("Stop[ return 0; return 1; return .i; ]", stop.toString());
-        assertEquals(2L, Evaluator.evaluate(stop,  3));
-    }
-
-    @Test
-    public void testCoRecur() {
-        Parser parser = new Parser(
-"""
-struct int0 { int i; flt0? f; }
-struct flt0 { flt f; int0? i; }
-int0 i0 = new int0;
-i0.i = 17;
-flt0 f0 = new flt0;
-f0.f = 3.14;
-i0.f = f0;
-f0.i = i0;
-return f0.i.f.i.i;
-""");
-        StopNode stop = parser.parse(false).iterate(false);
-        assertEquals("return 17;", stop.toString());
-    }
-
-    @Test
-    public void testNullRef0() {
-        Parser parser = new Parser(
-"""
-struct N { N next; int i; }
-N n = new N;
-return n.next;
-""");
-        StopNode stop = parser.parse(false).iterate(false);
-        assertEquals("return null;", stop.toString());
-    }
-
-    @Test
-    public void testNullRef1() {
-        Parser parser = new Parser(
-"""
-struct N { N next; int i; }
-N n = new N;
-n.next = new N;
-return n.next;
-""");
-        StopNode stop = parser.parse(false).iterate(false);
-        assertEquals("return new N;", stop.toString());
-    }
-
-    @Test
-    public void testNullRef2() {
-        Parser parser = new Parser(
-"""
-struct N { N next; int i; }
-N n = new N;
-n.next = null;
-return n.next;
-""");
-        try { parser.parse(true).iterate(true); fail(); }
-        catch( Exception e ) { assertEquals("Cannot store null into field *N next",e.getMessage()); }
-    }
-
-    @Test
-    public void testNullRef3() {
-        Parser parser = new Parser(
-"""
-struct N { N next; int i; }
-N n = new N;
-n.i = 3.14;
-return n.i;
-""");
-        try { parser.parse(true).iterate(true); fail(); }
-        catch( Exception e ) { assertEquals("Cannot store 3.14 into field int i",e.getMessage()); }
-    }
 }
